refactor(schedule): extract exec tools into registerScheduleExecTools (phase 6a)
Pulls the ten ctx.registerExecTool calls (list / create / update / pause /
delete / get / run_now / runs / parse / briefing) out of the activate() body
into lib/exec-tools.ts as registerScheduleExecTools(ctx). activate() now just
calls it. The handlers delegate to the job-service verbs, the fire engine
(fireManualRun), and the cron parser — all already extracted.

Pure behavior-preserving relocation — the tool definitions are verbatim.
919 -> 609.

Verified: typecheck + eslint + full schedule suite (280/0; routes.test.ts
carries the exec-tool callTool coverage) + full suite (5124/0) + binary build
+ rig: the schedule plugin activates against real OpenClaw, i.e. all ten
registrations run without throwing (a registration error would surface as
activation_failed). The underlying verbs were already E2E-proven on the live
stack in the fire-engine + job-service phases (run-now, automatic tick, CRUD).
